move table keys to new expression
