+++ packages/core/src/abap/2_statements/expressions/index.ts
@@ -179,6 +179,7 @@ export * from "./throw";
 export * from "./type_name_or_infer";
 export * from "./type_name";
 export * from "./type_param";
+export * from "./type_table_keys";
 export * from "./type_table";
 export * from "./type";
 export * from "./value_body";
