+++ packages/core/src/abap/2_statements/expressions/type_table.ts
@@ -1,25 +1,13 @@
-import {seq, opt, alt, ver, per, Expression, altPrio, plus, plusPrio, optPrio} from "../combi";
-import {Constant, FieldSub, TypeName, Integer, Field} from ".";
-import {Version} from "../../../version";
+import {seq, opt, alt, per, Expression, altPrio, optPrio} from "../combi";
+import {Constant, FieldSub, TypeName, Integer} from ".";
 import {IStatementRunnable} from "../statement_runnable";
 import {FieldChain} from "./field_chain";
+import {TypeTableKeys} from "./type_table_keys";
 
 export class TypeTable extends Expression {
   public getRunnable(): IStatementRunnable {
     const header = "WITH HEADER LINE";
     const initial = seq("INITIAL SIZE", Constant);
-
-    const uniqueness = alt("NON-UNIQUE", "UNIQUE");
-    const defaultKey = "DEFAULT KEY";
-    const emptyKey = ver(Version.v740sp02, "EMPTY KEY");
-
-    const key = seq("WITH",
-                    opt(uniqueness),
-                    altPrio(defaultKey, emptyKey,
-                            seq(opt(alt("SORTED", "HASHED")),
-                                "KEY",
-                                alt(seq(Field, "COMPONENTS", plus(FieldSub)),
-                                    plus(FieldSub)))));
 
     const normal1 = seq(opt(alt("STANDARD", "HASHED", "INDEX", "SORTED", "ANY")),
                         "TABLE",
@@ -27,19 +15,17 @@ export class TypeTable extends Expression {
                         opt("REF TO"),
                         opt(TypeName));
 
-    const keys = plusPrio(key);
-
     const likeType = seq(opt(alt("STANDARD", "HASHED", "INDEX", "SORTED", "ANY")),
                          "TABLE OF",
                          optPrio("REF TO"),
                          opt(FieldChain),
-                         opt(per(header, initial, keys)));
+                         opt(per(header, initial, TypeTableKeys)));
 
     const rangeType = seq("RANGE OF", TypeName, opt(header), opt(initial));
     const rangeLike = seq("RANGE OF", FieldSub, opt(header), opt(initial));
 
     const typetable = seq(normal1,
-                          opt(per(header, initial, keys)));
+                          opt(per(header, initial, TypeTableKeys)));
 
     const occurs = seq("OCCURS", Integer);
