+++ packages/core/src/rules/avoid_use.ts
@@ -2,7 +2,7 @@ import * as Statements from "../abap/2_statements/statements";
 import {Issue} from "../issue";
 import {ABAPRule} from "./_abap_rule";
 import {BasicRuleConfig} from "./_basic_rule_config";
-import {TypeTable} from "../abap/2_statements/expressions";
+import {TypeTable, TypeTableKeys} from "../abap/2_statements/expressions";
 import {IRuleMetadata, RuleTag} from "./_irule";
 import {ABAPFile} from "../abap/abap_file";
 import {Version} from "../version";
@@ -88,10 +88,9 @@ DESCRIBE TABLE LINES: use lines() instead (quickfix exists)`,
           && (this.reg.getConfig().getVersion() >= Version.v740sp02
           || this.reg.getConfig().getVersion() === Version.Cloud)
           && (statement instanceof Statements.Data || statement instanceof Statements.Type)) {
-        const tt = statementNode.findFirstExpression(TypeTable);
+        const tt = statementNode.findFirstExpression(TypeTable)?.findDirectExpression(TypeTableKeys);
         const token = tt?.findDirectTokenByText("DEFAULT");
         if (tt && token) {
-          tt.concatTokensWithoutStringsAndComments().toUpperCase().endsWith("DEFAULT KEY");
           message = "DEFAULT KEY";
           issues.push(Issue.atToken(file, token, this.getDescription(message), this.getMetadata().key, this.conf.severity));
         }
